LayoutTests/fast/scrolling/mac/scrollable-area-size-for-overlay-scrollbars.html

<!DOCTYPE html> <!-- webkit-test-runner [ AsyncOverflowScrollingEnabled=true MockScrollbarsEnabled=false ] -->
<html>
<head>
    <style>
        .container {
            height: 200px;
            width: 200px;
            border: 1px solid black;
            overflow-y: scroll;
        }
    </style>
    <script src="../../../resources/js-test.js"></script>
    <script src="../../../resources/ui-helper.js"></script>
    <script>
        if (window.testRunner)
            testRunner.waitUntilDone();
        description("Tests that scrollable area size is properly updated when toggling scrollbar style");
        window.internals.setUsesOverlayScrollbars(true);

        async function doTest()
        {
            preChangeSize = window.internals.scrollableAreaWidth(document.getElementById("scroller"));
            window.internals.setUsesOverlayScrollbars(false);
            shouldBeGreaterThan("preChangeSize", "window.internals.scrollableAreaWidth(document.getElementById(\"scroller\"))");
            testRunner.notifyDone();
        }
        window.addEventListener('load', doTest, false);
    </script>
</head>
<body>
    <div id="scroller" class="container">
            <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. .</p>

    </div>
</body>
</html>
