LayoutTests/fast/scrolling/mac/scrollable-area-size-for-overlay-scrollbars-expected.txt

Tests that scrollable area size is properly updated when toggling scrollbar style

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS preChangeSize is > window.internals.scrollableAreaWidth(document.getElementById("scroller"))
PASS successfullyParsed is true

TEST COMPLETE
Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. .
